Show plots from collection editor in Plots plugin
Depending on whether the Plots plugin is enabled and the user's
preferences, show plots from the collection editor either in a
separate window (the current behaviour) or in the Plots plugin.

In the latter case, emit a `sig_show_figure_requested` signal.
This signal propagates to the Variable Explorer plugin, which
calls the new `add_plot()` function in the Plots plugin.

Add an integration test to check that everything is coupled together.