Split ArrayEditor.setup_and_check() in two
Make two functions. The first is setup_ui(), which sets up the UI
and should be called only once. The second is set_data_and_check(),
which specifies the data for the editor and can be called repeatedly.
This is to prepare for the refresh functionality in the next commit.