Move incorrect codesign before set-key-partition-list.
This seems to correct the issue where codesign would hang while replacing signature for conda.exe.
I suspect that a gui prompt is launched, waiting for user input and giving the appearance of hanging.
I think this may have occurred because set-key-partition-list may have registered the wrong codesign and that is why the incorrect codesign must be moved before set-key-partition-list is called.