Set keychain logout timeout.
This also seems to fix the issue of hanging during codesigning conda.exe.
I suppose the hypothesis here is that the keychain locks before the signing event and a gui likely appears prompting the user for access.

The fact that both this and the location of set-key-partition-list appear to resolve the issue _independently_ (I tried each commit in isolation) is strange.
I could believe the keychain locks before signing, but then that may be intermittent: sometimes it locks sooner than others, but shouldn't it be required that set-key-partition-list be set after the incorrect codesign is moved?
In other words, why do these work independently? Shouldn't they each be necessary but insufficient?

Also, I still don't understand how napari avoids this issue.
They do not use set-key-partition-list at all, just the timeout.