Add `register` parameter to SpyderWidgetMixin.create_xxx()
The new parameter is added to the funcions create_menu(), create_toolbar()
and create_toolbutton. It indicates whether the menu, toolbar or
toolbuttion is added to the global registry. This is to allow the caller
to prevent adding temporary menus, toolbars and toolbuttons, such as those
in dialog boxes, to the registry forever.