Use SpyderWidgetMixin.create_toolbar in Variable Explorer editors
Use the function create_toolbar from SpyderWidgetMixin to create toolbars
and toolbuttons in the array editor, the collections editor, the dataframe
editor and the object explorer. This is to standardize the appearance and
the code. The biggest change is in the object explorer, where the toolbar
is clearly styled differently (and consistent with the rest of Spyder).