Use SpyderWidgetMixin.create_action in Variable Explorer editors
Use the function create_action from SpyderWidgetMixin to create actions
in the array editor, the collections editor, the dataframe editor and
the object explorer. This is to standardize the UX and the code.

This does remove the keyboard shortcuts for the actions in the object
explorer, but these shortcuts were not shown anywhere so users had
little change to discover them. This means that users are unlikely to
miss the shortcuts.