Move Refresh and Resize to front of toolbars
This commit re-orders the buttons in the toolbars for the array editor,
collections editor, dataframe editor and object explorer. The Refresh
button is the only button present for every editor and is probably one
of the most common buttons users will click, as well as being something
that applies to the whole object and always enabled, so it makes sense
to put it first. The Resize buttons (only present for the array and
dataframe editors) also apply to the whole object and are always
enabled, so they are put next.