Create DataFrameView only once in dataframe editor
Change DataFrameEditor so that it creates the DataFrameView in
setup_ui() and the DataFrameModel in set_data_and_check(). This
means that the DataFrameView is created only once, but the
DataFrameModel is created every time the data is refreshed. Before,
a new DataFrameView and a new DataFrameModel were created every time
the data was refreshed.

This fixes a bug where the scroll bars were not visible if the data
frame too big. This commit also adds a test for this.