Change default value of new parameter in create_toolbutton()
This fixes a mistake in commit 7d76a884. That commit introduced a
new parameter `register` in the member function create_toolbutton()
in SpyderToolButtonMixin. The default value made this a change in
behaviour which broke some tests of the outline explorer. The new
default value ensures that the behaviour is not changed if no value
for the new parameter is given, making the change backward
compatible.