Use UTF-8 by default for Python files per PEP3120
This PR changes how Spyder determines the encoding of files opened
in the editor to ensure that we follow PEP3120.

For Python files without a declaration of the form "coding=xxx",
use UTF-8 as encoding. For other files, we use the chardet library
as before.

For the purpose of this PR, Python files are those with one of
the following extensions: .py, .pyw or .ipy .

Also add tests for the new functionality.