Get rows/cols loaded in DataFrameHeaderModel from DataFrameModel
When setting the number of rows and columns loaded in the
DataFrameHeaderModel, use the information in the corresponding
DataFrameModel. This allows for the possibility that rows and
columns were loaded before the DataFrameHeaderModel was created.

This fixes a bug where a data frame editor with a large data
frame was opened, so that not all columns are initially loaded.
Then if you scroll to the last column and sort that column, a new
DataFrameHeaderModel was created which used the wrong value of
cols_loaded, causing the editor to scroll back. Also add a
regression test for the bug.