Change default format in array editor to ''
Before, the default format in the array editor for non-numeric arrays
was set to 's'. This meant that ArrayModel.data() returned a value of
type numpy.str_ for string arrays instead of a value of type str. The
consequence was that the contents of the array were not shown. Changing
the default format to '' fixes this.

Also add a regression test.