Allow files to be opened in another plugin than the editor
Any plugin can declare file extensions that it can open in the
FILE_EXTENSIONS attribute. If the user opens a file with one of
these extensions, then the open_file method in the plugin is
called. If no plugin is found then the file is opened in the
Editor plugin.

Extend the open_file test to cover this.