Move "New file" action from Editor to Application plugin
We want to generalize this action (and other actions in the File menu)
so that it can also be used to create new files in other panes, like the
Notebook pane. This commit is preparation and it only moves the code;
upcoming commit will allow other plugins to hook into the "New file" action.

This also moves the shortcut. The Application plugin owns the shortcut, but
does not expose it to the user. Plugins like the Editor plugin that want to
expose the shortcut need to do that themselves. This is because making the
"New file" and similar file shortcuts global would be confusing,
especially since Ctrl-W for "Close file" is already used in other plugins.

Also add a test for the action in the Application plugin.