Keep track of focused plugin in Application plugin
In the main window, keep track of the plugin that has keyboard
focus and emit sig_focused_plugin_changed to all plugin when
this changes.

Use this signal in the Application plugin to also keep track
of this information. Add a test for this.

This will be used in the following commit to dispatch file
operations to the appropriate plugin.