Allow other plugins to enable or disable file actions
The Application plugin is in charge of whether file actions are
enabled. Other plugins use enable_file_action() to indicate whether
the action is enabled for them. The Application plugin stores this
information and uses it to enable or disable file actions when focus
switches to another plugin.

Add a test for this.