Fix issue where adding duplicate path and selecting "yes" to move it to the top of the user list actually placed it in the project path section.
Fix issue where adding duplicate path that is also the same as project path and selecting "yes" to move it to the top of the user list actually moved the project path item to the user list section.

Ensure that user path item is selectable even if same as project path item.