pyside6: Fix error in context menu of IPython console
Fix KeyError when context menu in IPython console is opened for the
second or subsequent time.

QActions created by SpyderActionMixin.create_action() are, by default,
saved in a WeakValueDictonary. The context menu in the IPython console
is cleared and filled whenever it is opened. An attempt to retrieve the
required actions from the WeakValueDictonary fails if there are no
(more) strong references. Make sure there is a strong reference to
actions used in the context menu.