Fix editing of large ints in Variable Explorer
Raise the maximum of digits for converting an int to a string
to 32,767. Since Python 3.11, converting int with too many digits
to a string throws an error to prevent denial-of-service attacks.
This leads to an error when editing a large int in the Variable
Explorer.

When the user tries to edit an int with more than 32,767 digits,
a dialog box is displayed explaining that the value is too large.

The number 32,767 was chosen because that is the maximum length
of the QLineEdit that is used to edit an int.