Do not scroll to bottom if last plot is not selected
Change the logic in the Plots pane so that when a plot is added,
the scrollbar at the right will scroll to the bottom if this is
the first plot in the current execution OR (the scrollbar is
currently at the bottom AND the last plot is selected).

Before, the scrollbar would scroll to be bottom if it is the
first plot OR the scrollbar is currently at the bottom.