pyside6: Fix Spyder restart after changing settings
Changing certain settings requires a restart of Spyder. Spyder prompts
the user to do so when the settings dialog is accepted. Fix an error
(see comment in code) which ultimately prevented Spyder from restarting.

Details: The error fixed in this comment resulted in the "Issue
reporter" dialog being displayed. The dialog (the whole application) was
not visible when the error occured as closing of the application was
already triggered. The dialog blocked Spyder from actually exiting.