IPython console: Fix a couple of errors with remote consoles
- Prevent error when closing Spyder and a remote console is open. In
that case we don't need to send a request to shutdown the remote kernel
because stopping the server takes care of that.
- Don't check for related clients for remote ones. That avoids showing a
message asking to close all related clients when closing a remote one,
which closes all if you click Yes and doesn't make sense because they
are all connected to different kernels.