remove additional usages and finally definition of add_cursor
Only one instance of add_cursor was not in a loop for all extra cursors which was found to be a slow pattern for many cursors. After refactoring those instances there was only one case where add_cursor was used, so the function was inlined there as well and the definition was removed.